refactor(agent): 析出 SDK options 装配为持依赖装配器，凭证注入与 hook 工厂随迁
SessionManager 不再直接构建 options 与 hook，仅通过 OptionsAssembler.build 装配；DB 凭证注入（Anthropic 真值 + 其他供应商空值覆盖，env_keys 仍为名单唯一真相源）、prompt 变体与项目上下文、各 PreToolUse/PostToolUse hook 工厂（含 JSON 校验）、AgentAccessPolicy 的 settings 编译与裁决 adapter 接线均归装配器。access_policy 与 max_turns 以 provider 回调现取，冷启动/重连/locale 重建行为不变。装配器测试改以注入假依赖驱动，既有断言逐字搬家。
