+++ tests/agent_runtime/test_agent_startup_error.py
@@ -65,10 +65,10 @@ async def test_build_options_forwards_stderr_callback(
 ) -> None:
     """_build_options 必须把 stderr 回调透传给 ClaudeAgentOptions。"""
 
-    async def fake_env(_self):
+    async def fake_env():
         return {"ANTHROPIC_API_KEY": "sk"}
 
-    monkeypatch.setattr(SessionManager, "_build_provider_env_overrides", fake_env)
+    monkeypatch.setattr("server.agent_runtime.options_assembler.load_provider_env_overrides", fake_env)
 
     captured: list[str] = []
 
@@ -89,10 +89,10 @@ async def test_send_new_session_wraps_actor_failure_with_stderr(
 ) -> None:
     """actor.start() 抛错时应收集 stderr 并包装为 AgentStartupError 抛出。"""
 
-    async def fake_env(_self):
+    async def fake_env():
         return {"ANTHROPIC_API_KEY": "sk"}
 
-    monkeypatch.setattr(SessionManager, "_build_provider_env_overrides", fake_env)
+    monkeypatch.setattr("server.agent_runtime.options_assembler.load_provider_env_overrides", fake_env)
 
     captured_stderr_cb: list = []
 
@@ -145,10 +145,10 @@ async def test_send_new_session_no_stderr_still_wraps(
 ) -> None:
     """即使 SDK 没产生 stderr，actor.start 失败也应包装为 AgentStartupError（保留原因链）。"""
 
-    async def fake_env(_self):
+    async def fake_env():
         return {"ANTHROPIC_API_KEY": "sk"}
 
-    monkeypatch.setattr(SessionManager, "_build_provider_env_overrides", fake_env)
+    monkeypatch.setattr("server.agent_runtime.options_assembler.load_provider_env_overrides", fake_env)
 
     class _FakeActor:
         def __init__(self, *_, on_message=None, client_factory=None):
@@ -183,10 +183,10 @@ async def test_get_or_connect_wraps_actor_failure_with_stderr(
     """
     from tests.factories import make_session_meta
 
-    async def fake_env(_self):
+    async def fake_env():
         return {"ANTHROPIC_API_KEY": "sk"}
 
-    monkeypatch.setattr(SessionManager, "_build_provider_env_overrides", fake_env)
+    monkeypatch.setattr("server.agent_runtime.options_assembler.load_provider_env_overrides", fake_env)
 
     captured_stderr_cb: list = []
 
@@ -235,10 +235,10 @@ async def test_stderr_buffer_caps_to_maxlen(session_manager: SessionManager, mon
     可控；这里直接打超过上限的行数，断言 sdk_stderr 行数受限。
     """
 
-    async def fake_env(_self):
+    async def fake_env():
         return {"ANTHROPIC_API_KEY": "sk"}
 
-    monkeypatch.setattr(SessionManager, "_build_provider_env_overrides", fake_env)
+    monkeypatch.setattr("server.agent_runtime.options_assembler.load_provider_env_overrides", fake_env)
 
     captured_stderr_cb: list = []
     overflow_count = _SDK_STDERR_BUFFER_MAX + 50
