+++ tests/agent_runtime/test_session_manager_store_injection.py
@@ -10,7 +10,7 @@ from lib.agent_session_store.store import DbSessionStore
 from server.agent_runtime.session_manager import SessionManager
 
 
-async def _fake_provider_env(_self):
+async def _fake_provider_env():
     """Stub: 跳过 DB 访问，返回空 dict（不影响 session_store/flush 字段断言）。"""
     return {}
 
@@ -57,13 +57,13 @@ def test_store_db_explicit_returns_store(monkeypatch, tmp_path):
 
 
 def test_store_uses_session_factory_seam(monkeypatch, tmp_path):
-    """If sm._session_factory is set, _build_session_store uses it."""
+    """装配器的 session_factory / user_id 注入生效于 build_session_store。"""
     monkeypatch.delenv("ARCREEL_SDK_SESSION_STORE", raising=False)
     sm = _build_sm(tmp_path)
 
     sentinel = object()
-    sm._session_factory = sentinel  # type: ignore[attr-defined]
-    sm._user_id = "test-user"  # type: ignore[attr-defined]
+    sm._options_assembler._session_factory = sentinel  # type: ignore[attr-defined]
+    sm._options_assembler._user_id = "test-user"  # type: ignore[attr-defined]
 
     store = sm._build_session_store()
     assert isinstance(store, DbSessionStore)
@@ -75,7 +75,7 @@ def test_store_uses_session_factory_seam(monkeypatch, tmp_path):
 async def test_flush_mode_passed_to_options_default(monkeypatch, tmp_path):
     """No env → ClaudeAgentOptions.session_store_flush == 'eager'."""
     monkeypatch.delenv("ARCREEL_SDK_SESSION_STORE_FLUSH", raising=False)
-    monkeypatch.setattr(SessionManager, "_build_provider_env_overrides", _fake_provider_env)
+    monkeypatch.setattr("server.agent_runtime.options_assembler.load_provider_env_overrides", _fake_provider_env)
     sm = _build_sm(tmp_path)
 
     project_cwd = tmp_path / "projects" / "demo"
@@ -89,7 +89,7 @@ async def test_flush_mode_passed_to_options_default(monkeypatch, tmp_path):
 async def test_flush_mode_passed_to_options_batched(monkeypatch, tmp_path):
     """env=batched → options.session_store_flush == 'batched'."""
     monkeypatch.setenv("ARCREEL_SDK_SESSION_STORE_FLUSH", "batched")
-    monkeypatch.setattr(SessionManager, "_build_provider_env_overrides", _fake_provider_env)
+    monkeypatch.setattr("server.agent_runtime.options_assembler.load_provider_env_overrides", _fake_provider_env)
     sm = _build_sm(tmp_path)
     project_cwd = tmp_path / "projects" / "demo"
     project_cwd.mkdir(parents=True)
@@ -108,7 +108,7 @@ async def test_flush_mode_passed_to_options_when_store_off(monkeypatch, tmp_path
     """
     monkeypatch.setenv("ARCREEL_SDK_SESSION_STORE", "off")
     monkeypatch.delenv("ARCREEL_SDK_SESSION_STORE_FLUSH", raising=False)
-    monkeypatch.setattr(SessionManager, "_build_provider_env_overrides", _fake_provider_env)
+    monkeypatch.setattr("server.agent_runtime.options_assembler.load_provider_env_overrides", _fake_provider_env)
     sm = _build_sm(tmp_path)
 
     project_cwd = tmp_path / "projects" / "demo"
